+++ debomatic-webui/routes/index.js
@@ -1,18 +1,19 @@
+'use strict';
 
-/*
- * GET home page.
- */
+var config = require('../lib/config.js');
 
-var config = require('../lib/config.js')
-
-exports.index = function(req, res){
-  res.render('index', config)
+exports.index = function (req, res) {
+    res.render('index', config);
 };
 
-exports.distribution = function(req, res) {
-  res.render('distribution', config)
+exports.distribution = function (req, res) {
+    res.render('distribution', config);
 };
 
-exports.preferences = function(req, res) {
-	res.render('preferences', config)
-}
+exports.preferences = function (req, res) {
+    res.render('preferences', config);
+};
+
+exports.commands = function (req, res) {
+    res.render('commands', config);
+};