+++ debomatic-webui/lib/config.js
@@ -30,9 +30,8 @@ config.routes.distribution = '/distribution';
 config.routes.preferences = '/preferences';
 config.routes.commands = '/commands';
 
+// web configuration
 config.web = {};
-config.web.title = 'Deb-o-Matic web.ui';
-config.web.description = 'This is a web interface for debomatic';
 
 // debomatic configuration exportable for web
 config.web.debomatic = {};
@@ -47,6 +46,10 @@ config.web.debomatic.dput.host = config.host;
 config.web.debomatic.dput.login = 'debomatic';
 config.web.debomatic.dput.method = 'scp';
 config.web.debomatic.dput.unsigned_uploads = false;
+
+// header title and description
+config.web.title = 'Deb-o-Matic web.ui';
+config.web.description = 'This is a web interface for debomatic over ' + config.web.debomatic.architecture;
 
 // list of files get preview
 config.web.file = {};