+++ History.md
@@ -5,6 +5,7 @@
  * [new] add page commands which documents the tasks interface of Deb-o-Matic
  * [new] better tooltip on action buttons (such as Download, View al file, Get all)
  * [new] focus debomatic architecture in header description
+ * [new] if header is hidden show architecture and version in footer
 
 # 0.4.2 (2014-06-15)
  * [fix] calculate pidfile for debomatic status check after merging configurations