+++ History.md
@@ -1,4 +1,4 @@
-# 0.5.0
+# 0.5.0 (2014-06-16)
  * [new] add packages search bar
  * [new] get buildlog preview by default instead of datestamp when click on a package
  * [new] always show datestamp information below the title
@@ -6,6 +6,7 @@
  * [new] better tooltip over the action buttons (such as Download, View al file, Get all)
  * [new] show the debomatic architecture in the header description
  * [new] show architecture and version in footer if header is hidden in preferences
+ * [fix] when watching a file which gets deletion tail lib crashes
 
 # 0.4.2 (2014-06-15)
  * [fix] calculate pidfile for debomatic status check after merging configurations